add: authorization in react

+++ frontend/src/components/WorkoutDetail.jsx
@@ -1,12 +1,19 @@
+import { useAuthContext } from "../hooks/useAuthContext";
 import { useWorkoutContext } from "../hooks/useWorkoutContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutDetail = ({ workout }) => {
   const { dispatch } = useWorkoutContext();
+  const { user } = useAuthContext();
 
   const onDeleteHandler = async () => {
+    if (!user) return;
+
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     await response.json();
     if (response.ok) dispatch({ type: "DELETE_WORKOUT", payload: workout._id });
